requirements.txt
streamlit==1.44.1
langchain==0.3.23
faiss-gpu==1.7.2
sentence-transformers==4.1.0
python-dotenv==1.1.0
llama-cpp-python==0.2.26
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
tqdm==4.66.1
requests==2.31.0